Improve logger initialization and traceback logging
Logging initialization is now done with the object LogSetup.
It enable fine grained initialization and provide few usefull methods.

Traceback logging is improved based on LogSetup
Tracebacks are properly handled and specially when no logging facility
has been initialized.

TODO: This change is a raw copy from ensime-common.
It would be great to unify both, but this imply creating a new addon or...?